+++ src/model/category.ts
@@ -2,7 +2,7 @@ import {Column, Entity, PrimaryGeneratedColumn} from "typeorm";
 
 @Entity({name: 'category'})
 export class Category {
-    @PrimaryGeneratedColumn({type: 'int'})
+    @PrimaryGeneratedColumn({type: 'bigint'})
     public readonly category_id: number
     @Column({type: 'varchar'})
     public category_name: string
