+++ src/model/exam-question.ts
@@ -2,10 +2,10 @@ import {Column, Entity, PrimaryGeneratedColumn} from "typeorm";
 
 @Entity({name: 'exam-question'})
 export class ExamQuestion {
-    @PrimaryGeneratedColumn({type: 'int'})
+    @PrimaryGeneratedColumn({type: 'bigint'})
     public readonly examQuestion_id: number
-    @Column({type: 'int'})
+    @Column({type: 'bigint'})
     public exam_id: number
-    @Column({type: 'int'})
+    @Column({type: 'bigint'})
     public question_id: number
 }
